requirements.txt
click>=8.1.7
requests>=2.31.0
pydantic>=2.7.0
python-dotenv>=1.0.1
rich>=13.7.1

# API Providers
openai>=1.14.0
anthropic>=0.23.0
google-generativeai>=0.4.1

# Dev / Testing
pytest>=8.1.1
pytest-cov>=5.0.0
ruff>=0.3.4